refactor mocks to be generalized and reusable
