+++ native/core/src/lib.rs
@@ -5,3 +5,6 @@ pub mod config;
 pub mod execution;
 pub mod meta;
 pub mod referencing;
+
+#[cfg(any(test, feature = "testing"))]
+pub mod mocks;
