+++ native/core/src/mocks/mod.rs
@@ -1,0 +1,2 @@
+mod store;
+pub use store::*;
